@@ -1,10 +1,9 @@
 ﻿using System.Threading.Tasks;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Core.Interfaces
+namespace Equinor.ProCoSys.BusSenderWorker.Core.Interfaces;
+
+public interface IEntryPointService
 {
-    public interface IEntryPointService
-    {
-        Task SendMessageChunk();
-        Task StopService();
-    }
+    Task SendMessageChunk();
+    Task StopService();
 }
