@@ -1,9 +1,8 @@
 ﻿using System.Threading.Tasks;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Core.Interfaces
+namespace Equinor.ProCoSys.BusSenderWorker.Core.Interfaces;
+
+public interface ITagDetailsRepository
 {
-    public interface ITagDetailsRepository
-    {
-        public Task<string> GetDetailsStringByTagId(long tagId);
-    }
+    public Task<string> GetDetailsStringByTagId(long tagId);
 }
