@@ -1,10 +1,9 @@
-﻿namespace Equinor.ProCoSys.BusSenderWorker.Core.Models
+﻿namespace Equinor.ProCoSys.BusSenderWorker.Core.Models;
+
+public class BusEventMessage
 {
-    public class BusEventMessage
-    {
-        public string Plant { get; set; }
-        public string ProjectName { get; set; }
-        public string McPkgNo { get; set; }
-        public string CommPkgNo { get; set; }
-    }
+    public string Plant { get; set; }
+    public string ProjectName { get; set; }
+    public string McPkgNo { get; set; }
+    public string CommPkgNo { get; set; }
 }
