@@ -1,10 +1,9 @@
 ﻿
 
-namespace Equinor.ProCoSys.BusSenderWorker.Core.Models
+namespace Equinor.ProCoSys.BusSenderWorker.Core.Models;
+
+public class WoMaterialIdentifier
 {
-    public class WoMaterialIdentifier
-    {
-        public string WoId { get; set; }
-        public string ItemNo { get; set;}
-    }
+    public string WoId { get; set; }
+    public string ItemNo { get; set;}
 }
