@@ -1,11 +1,10 @@
-﻿namespace Equinor.ProCoSys.BusSenderWorker.Core
+﻿namespace Equinor.ProCoSys.BusSenderWorker.Core;
+
+public enum Status
 {
-    public enum Status
-    {
-        UnProcessed = 0,
-        Processing = 1,
-        Failed = 2,
-        Sent = 3,
-        Skipped = 4
-    }
+    UnProcessed = 0,
+    Processing = 1,
+    Failed = 2,
+    Sent = 3,
+    Skipped = 4
 }
