@@ -3,45 +3,44 @@ using System.Collections.Generic;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Extensibility;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Core.Telemetry
+namespace Equinor.ProCoSys.BusSenderWorker.Core.Telemetry;
+
+public class ApplicationInsightsTelemetryClient : ITelemetryClient
 {
-    public class ApplicationInsightsTelemetryClient : ITelemetryClient
+    private readonly TelemetryClient _ai;
+
+    public ApplicationInsightsTelemetryClient(TelemetryConfiguration telemetryConfiguration)
     {
-        private readonly TelemetryClient _ai;
-
-        public ApplicationInsightsTelemetryClient(TelemetryConfiguration telemetryConfiguration)
+        if (telemetryConfiguration == null)
         {
-            if (telemetryConfiguration == null)
-            {
-                throw new ArgumentNullException(nameof(telemetryConfiguration));
-            }
-
-            _ai = new TelemetryClient(telemetryConfiguration)
-            {
-                // The InstrumentationKey isn't set through the configuration object. Setting it explicitly works.
-                InstrumentationKey = telemetryConfiguration.InstrumentationKey
-            };
+            throw new ArgumentNullException(nameof(telemetryConfiguration));
         }
 
-        public void TrackEvent(string name, Dictionary<string, string> properties) =>
-            _ai
-                .TrackEvent(name, properties);
-
-        public void TrackMetric(string name, double metric) =>
-            _ai
-                .GetMetric(name)
-                .TrackValue(metric);
-
-        public void TrackMetric(string name, double metric, string dimension1Name, string dimension1Value) =>
-            _ai
-                .GetMetric(name, dimension1Name)
-                .TrackValue(metric, dimension1Value);
-
-        public void TrackMetric(string name, double metric, string dimension1Name, string dimension2Name, string dimension1Value, string dimension2Value) =>
-            _ai
-                .GetMetric(name, dimension1Name, dimension2Name)
-                .TrackValue(metric, dimension1Value, dimension2Value);
-
-        public void Flush() => _ai.Flush();
+        _ai = new TelemetryClient(telemetryConfiguration)
+        {
+            // The InstrumentationKey isn't set through the configuration object. Setting it explicitly works.
+            InstrumentationKey = telemetryConfiguration.InstrumentationKey
+        };
     }
+
+    public void TrackEvent(string name, Dictionary<string, string> properties) =>
+        _ai
+            .TrackEvent(name, properties);
+
+    public void TrackMetric(string name, double metric) =>
+        _ai
+            .GetMetric(name)
+            .TrackValue(metric);
+
+    public void TrackMetric(string name, double metric, string dimension1Name, string dimension1Value) =>
+        _ai
+            .GetMetric(name, dimension1Name)
+            .TrackValue(metric, dimension1Value);
+
+    public void TrackMetric(string name, double metric, string dimension1Name, string dimension2Name, string dimension1Value, string dimension2Value) =>
+        _ai
+            .GetMetric(name, dimension1Name, dimension2Name)
+            .TrackValue(metric, dimension1Value, dimension2Value);
+
+    public void Flush() => _ai.Flush();
 }
