@@ -4,16 +4,15 @@ using Equinor.ProCoSys.BusSenderWorker.Core.Models;
 using Equinor.ProCoSys.BusSenderWorker.Infrastructure.EntityConfiguration;
 using Microsoft.EntityFrameworkCore;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Data
+namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Data;
+
+public class BusSenderServiceContext : DbContext, IUnitOfWork
 {
-    public class BusSenderServiceContext : DbContext, IUnitOfWork
-    {
-        public BusSenderServiceContext(DbContextOptions<BusSenderServiceContext> options) : base(options) { }
+    public BusSenderServiceContext(DbContextOptions<BusSenderServiceContext> options) : base(options) { }
 
-        public virtual DbSet<BusEvent> BusEvents { get; set; } = null!;
-        public Task<int> SaveChangesAsync() => base.SaveChangesAsync();
+    public virtual DbSet<BusEvent> BusEvents { get; set; } = null!;
+    public Task<int> SaveChangesAsync() => base.SaveChangesAsync();
 
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder) => modelBuilder.ApplyConfiguration(new BusEventConfiguration());
-    }
+    protected override void OnModelCreating(ModelBuilder modelBuilder) => modelBuilder.ApplyConfiguration(new BusEventConfiguration());
 }
