@@ -1,21 +1,20 @@
 ﻿using Azure.Storage.Blobs;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Repositories
+namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Repositories;
+
+public class BlobRepository
 {
-    public class BlobRepository
+    private readonly BlobContainerClient _client;
+
+    public BlobRepository(string connectionString, string containerName) 
+        => _client = new BlobContainerClient(connectionString, containerName);
+
+    public async void Download(string pathAndFileName, string downloadPath)
     {
-        private readonly BlobContainerClient _client;
-
-        public BlobRepository(string connectionString, string containerName) 
-            => _client = new BlobContainerClient(connectionString, containerName);
-
-        public async void Download(string pathAndFileName, string downloadPath)
+        var blobClient = _client.GetBlobClient(pathAndFileName);
+        if (await blobClient.ExistsAsync())
         {
-            var blobClient = _client.GetBlobClient(pathAndFileName);
-            if (await blobClient.ExistsAsync())
-            {
-                await blobClient.DownloadToAsync(downloadPath);
-            }
+            await blobClient.DownloadToAsync(downloadPath);
         }
     }
 }
