@@ -8,24 +8,23 @@ using Equinor.ProCoSys.BusSenderWorker.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Repositories
+namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Repositories;
+
+public class BusEventRepository : IBusEventRepository
 {
-    public class BusEventRepository : IBusEventRepository
+    private readonly int _messageChunkSize;
+    private readonly DbSet<BusEvent> _busEvents;
+
+    public BusEventRepository(BusSenderServiceContext context, IConfiguration configuration)
     {
-        private readonly int _messageChunkSize;
-        private readonly DbSet<BusEvent> _busEvents;
+        _messageChunkSize = int.Parse(configuration["MessageChunkSize"]);
+        _busEvents = context.BusEvents;
+    }
 
-        public BusEventRepository(BusSenderServiceContext context, IConfiguration configuration)
-        {
-            _messageChunkSize = int.Parse(configuration["MessageChunkSize"]);
-            _busEvents = context.BusEvents;
-        }
-
-        public async Task<List<BusEvent>> GetEarliestUnProcessedEventChunk()
-        {
-            var events = await _busEvents.Where(e => e.Sent == Status.UnProcessed).OrderBy(e => e.Created).Take(_messageChunkSize).ToListAsync();
+    public async Task<List<BusEvent>> GetEarliestUnProcessedEventChunk()
+    {
+        var events = await _busEvents.Where(e => e.Sent == Status.UnProcessed).OrderBy(e => e.Created).Take(_messageChunkSize).ToListAsync();
             
-            return events;
-        }
+        return events;
     }
 }
