@@ -10,123 +10,125 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace Equinor.ProCoSys.BusSender.Worker
+namespace Equinor.ProCoSys.BusSender.Worker;
+
+public class Program
 {
-    public class Program
+    public IConfiguration Configuration { get; }
+
+    public Program(IConfiguration configuration) 
+        => Configuration = configuration;
+
+    public static async Task Main(string[] args)
     {
-        public IConfiguration Configuration { get; }
+        using var host = CreateHostBuilder(args).Build();
+        ILogger? logger = host.Services.GetService<ILogger<Program>>();
+        await host.RunAsync();
+    }
 
-        public Program(IConfiguration configuration) 
-            => Configuration = configuration;
-
-        public static async Task Main(string[] args)
-        {
-            using var host = CreateHostBuilder(args).Build();
-            ILogger? logger = host.Services.GetService<ILogger<Program>>();
-            await host.RunAsync();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args)
-        {
-            var builder = Host.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((context, config) =>
-                {
-                    config.AddUserSecrets<Program>(true);
-                    var settings = config.Build();
-
-                    var azConfig = settings.GetValue<bool>("UseAzureAppConfiguration");
-                    if (azConfig)
-                    {
-                        config.AddAzureAppConfiguration(options =>
-                        {
-                            var connectionString = settings["ConnectionStrings:AppConfig"];
-                            options.Connect(connectionString)
-                                .ConfigureKeyVault(kv =>
-                                {
-                                    kv.SetCredential(new DefaultAzureCredential());
-                                })
-                                .Select(KeyFilter.Any)
-                                .Select(KeyFilter.Any, context.HostingEnvironment.EnvironmentName)
-                                .ConfigureRefresh(refreshOptions =>
-                                {
-                                    refreshOptions.Register("Sentinel", true);
-                                    refreshOptions.SetCacheExpiration(TimeSpan.FromMinutes(5));
-                                });
-                        });
-                    }
-
-                    else if (settings["EnvironmentName"] == "Local")
-                    {
-                        config.AddAzureAppConfiguration(options =>
-                        {
-
-                            options.Connect(settings["AppConfig"])
-                                .ConfigureKeyVault(kv =>
-                                {
-                                    kv.SetCredential(new DefaultAzureCredential());
-                                })
-                                .ConfigureRefresh(options =>
-                                {
-                                    options.Register("Sentinel", true);
-                                    options.SetCacheExpiration(TimeSpan.FromMinutes(5));
-                                })
-                                .Select(KeyFilter.Any, LabelFilter.Null)
-                                .Select(KeyFilter.Any, settings["Azure:AppConfigLabelFilter"]);
-                        });
-                    }
-                    else
-                    {
-                        config.AddAzureAppConfiguration(options =>
-                        {
-                            var appConfigurationName = settings["Azure:AppConfig"];
-                            var endpoint = $"https://{appConfigurationName}.azconfig.io";
-
-                            options.Connect(new Uri(endpoint), new DefaultAzureCredential())
-                                .ConfigureKeyVault(kv =>
-                                {
-                                    kv.SetCredential(new DefaultAzureCredential());
-                                })
-                                .ConfigureRefresh(options =>
-                                {
-                                    options.Register("Sentinel", true);
-                                    options.SetCacheExpiration(TimeSpan.FromMinutes(5));
-                                })
-                                .Select(KeyFilter.Any, LabelFilter.Null)
-                                .Select(KeyFilter.Any, settings["Azure:AppConfigLabelFilter"]);
-                        });
-                    }
-                });
-
-            builder.ConfigureLogging((context, logging) =>
+    public static IHostBuilder CreateHostBuilder(string[] args)
+    {
+        var builder = Host.CreateDefaultBuilder(args)
+            .ConfigureAppConfiguration((context, config) =>
             {
-                logging.ClearProviders();
-                logging.AddConsole();
-                logging.AddApplicationInsightsWebJobs(c => c.InstrumentationKey = context.Configuration["ApplicationInsights:InstrumentationKey"]);
+                config.AddUserSecrets<Program>(true);
+                var settings = config.Build();
+
+                var azConfig = settings.GetValue<bool>("UseAzureAppConfiguration");
+                if (azConfig)
+                {
+                    config.AddAzureAppConfiguration(options =>
+                    {
+                        var connectionString = settings["ConnectionStrings:AppConfig"];
+                        options.Connect(connectionString)
+                            .ConfigureKeyVault(kv =>
+                            {
+                                kv.SetCredential(new DefaultAzureCredential());
+                            })
+                            .Select(KeyFilter.Any)
+                            .Select(KeyFilter.Any, context.HostingEnvironment.EnvironmentName)
+                            .ConfigureRefresh(refreshOptions =>
+                            {
+                                refreshOptions.Register("Sentinel", true);
+                                refreshOptions.SetCacheExpiration(TimeSpan.FromMinutes(5));
+                            });
+                    });
+                }
+
+                else if (settings["EnvironmentName"] == "Local")
+                {
+                    config.AddAzureAppConfiguration(options =>
+                    {
+
+                        options.Connect(settings["AppConfig"])
+                            .ConfigureKeyVault(kv =>
+                            {
+                                kv.SetCredential(new DefaultAzureCredential());
+                            })
+                            .ConfigureRefresh(options =>
+                            {
+                                options.Register("Sentinel", true);
+                                options.SetCacheExpiration(TimeSpan.FromMinutes(5));
+                            })
+                            .Select(KeyFilter.Any, LabelFilter.Null)
+                            .Select(KeyFilter.Any, settings["Azure:AppConfigLabelFilter"]);
+                    });
+                }
+                else
+                {
+                    config.AddAzureAppConfiguration(options =>
+                    {
+                        var appConfigurationName = settings["Azure:AppConfig"];
+                        var endpoint = $"https://{appConfigurationName}.azconfig.io";
+
+                        options.Connect(new Uri(endpoint), new DefaultAzureCredential())
+                            .ConfigureKeyVault(kv =>
+                            {
+                                kv.SetCredential(new DefaultAzureCredential());
+                            })
+                            .ConfigureRefresh(options =>
+                            {
+                                options.Register("Sentinel", true);
+                                options.SetCacheExpiration(TimeSpan.FromMinutes(5));
+                            })
+                            .Select(KeyFilter.Any, LabelFilter.Null)
+                            .Select(KeyFilter.Any, settings["Azure:AppConfigLabelFilter"]);
+                    });
+                }
             });
 
-            builder.UseContentRoot(Directory.GetCurrentDirectory())
-           .ConfigureServices((hostContext, services) =>
-           {
-               var rep = new BlobRepository(hostContext.Configuration["BlobStorage:ConnectionString"], hostContext.Configuration["BlobStorage:ContainerName"]);
+        builder.ConfigureLogging((context, logging) =>
+        {
+            logging.ClearProviders();
+            logging.AddConsole();
+            logging.AddApplicationInsightsWebJobs(c => c.InstrumentationKey = context.Configuration["ApplicationInsights:InstrumentationKey"]);
+        });
 
-               var walletPath = hostContext.Configuration["WalletFileDir"];
-               Directory.CreateDirectory(walletPath);
+        builder.UseContentRoot(Directory.GetCurrentDirectory())
+            .ConfigureServices((hostContext, services) =>
+            {
+                var rep = new BlobRepository(hostContext.Configuration["BlobStorage:ConnectionString"], hostContext.Configuration["BlobStorage:ContainerName"]);
 
-               rep.Download(hostContext.Configuration["BlobStorage:WalletFileName"], walletPath + "\\cwallet.sso");
-               Console.WriteLine("Created wallet file at: " + walletPath);
+                var walletPath = hostContext.Configuration["WalletFileDir"];
+                Directory.CreateDirectory(walletPath);
 
-               services.AddApplicationInsightsTelemetryWorkerService(hostContext.Configuration["ApplicationInsights:InstrumentationKey"]);
-               services.AddDbContext(hostContext.Configuration["ConnectionString"]);
-               services.AddTopicClients(
-                   hostContext.Configuration["ServiceBusConnectionString"],
-                   hostContext.Configuration["TopicNames"]);
-               services.AddRepositories();
-               services.AddServices();
+                rep.Download(hostContext.Configuration["BlobStorage:WalletFileName"], walletPath + "\\cwallet.sso");
+                Console.WriteLine("Created wallet file at: " + walletPath);
 
-               services.AddHostedService<TimedWorkerService>();
-           });
+                services.AddApplicationInsightsTelemetryWorkerService(hostContext.Configuration["ApplicationInsights:InstrumentationKey"]);
+                
+                var connectionString = hostContext.Configuration["ConnectionString"];
+                //var localConnectionString = hostContext.Configuration["ProcosysDb"]; //to be used when debugging
+                services.AddDbContext(connectionString);
+                services.AddTopicClients(
+                    hostContext.Configuration["ServiceBusConnectionString"],
+                    hostContext.Configuration["TopicNames"]);
+                services.AddRepositories();
+                services.AddServices();
 
-            return builder;
-        }
+                services.AddHostedService<TimedWorkerService>();
+            });
+
+        return builder;
     }
 }
