@@ -1,24 +1,23 @@
-﻿namespace Equinor.ProCoSys.PcsServiceBus
+﻿namespace Equinor.ProCoSys.PcsServiceBus;
+
+public class PcsServiceBusTelemetryConstants
 {
-    public class PcsServiceBusTelemetryConstants
-    {
-        public const string Event = "Event";
+    public const string Event = "Event";
 
-        public const string Plant = "Plant";
+    public const string Plant = "Plant";
 
-        public const string ProjectName = "ProjectName";
-        public const string ProjectNameOld = "ProjectNameOld";
-        public const string McPkgNo = "McPkgNo";
-        public const string McPkgNoOld = "McPkgNoOld";
-        public const string CommPkgNo = "CommPkgNo";
-        public const string Ipo = "Ipo";
-        public const string IpoEvent = "IpoEvent";
-        public const string CommPkgNoOld = "CommPkgNoOld";
-        public const string ResponsibleCode = "ResponsibleCode";
-        public const string IsVoided = "IsVoided";
-        public const string Code = "Code";
-        public const string RegisterCode = "RegisterCode";
-        public const string TagNo = "TagNo";
-        public const string TagNoOld = "TagNoOld";
-    }
+    public const string ProjectName = "ProjectName";
+    public const string ProjectNameOld = "ProjectNameOld";
+    public const string McPkgNo = "McPkgNo";
+    public const string McPkgNoOld = "McPkgNoOld";
+    public const string CommPkgNo = "CommPkgNo";
+    public const string Ipo = "Ipo";
+    public const string IpoEvent = "IpoEvent";
+    public const string CommPkgNoOld = "CommPkgNoOld";
+    public const string ResponsibleCode = "ResponsibleCode";
+    public const string IsVoided = "IsVoided";
+    public const string Code = "Code";
+    public const string RegisterCode = "RegisterCode";
+    public const string TagNo = "TagNo";
+    public const string TagNoOld = "TagNoOld";
 }
