@@ -1,39 +1,38 @@
-﻿namespace Equinor.ProCoSys.PcsServiceBus
+﻿namespace Equinor.ProCoSys.PcsServiceBus;
+
+public enum PcsTopic
 {
-    public enum PcsTopic
-    {
-        CommPkg,
-        Ipo,
-        McPkg,
-        Project,
-        Responsible,
-        Tag,
-        TagFunction,
-        PunchListItem,
-        Library,
-        WorkOrder,
-        Checklist,
-        Milestone,
-        WorkOrderCutoff,
-        Certificate,
-        WoChecklist,
-        SWCR,
-        SWCRSignature,
-        PipingRevision,
-        Tag_Delete,
-        WoChecklist_Delete,
-        WoMaterial_Delete,
-        PunchListItem_Delete,
-        PipingRevision_Delete,
-        WoMaterial,
-        WoMilestone,
-        Stock,
-        Document,
-        CommPkgOperation,
-        PipingSpool,
-        PipingSpool_Delete,
-        CommPkg_Delete,
-        SWCR_Delete,
-        McPkg_Delete
-    }
+    CommPkg,
+    Ipo,
+    McPkg,
+    Project,
+    Responsible,
+    Tag,
+    TagFunction,
+    PunchListItem,
+    Library,
+    WorkOrder,
+    Checklist,
+    Milestone,
+    WorkOrderCutoff,
+    Certificate,
+    WoChecklist,
+    SWCR,
+    SWCRSignature,
+    PipingRevision,
+    Tag_Delete,
+    WoChecklist_Delete,
+    WoMaterial_Delete,
+    PunchListItem_Delete,
+    PipingRevision_Delete,
+    WoMaterial,
+    WoMilestone,
+    Stock,
+    Document,
+    CommPkgOperation,
+    PipingSpool,
+    PipingSpool_Delete,
+    CommPkg_Delete,
+    SWCR_Delete,
+    McPkg_Delete
 }
