@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace Equinor.ProCoSys.PcsServiceBus.Receiver.Interfaces
+namespace Equinor.ProCoSys.PcsServiceBus.Receiver.Interfaces;
+
+public interface ILeaderElectorService
 {
-    public interface ILeaderElectorService
-    {
-        Task<bool> CanProceedAsLeader(Guid id);
-    }
+    Task<bool> CanProceedAsLeader(Guid id);
 }
