@@ -3,13 +3,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Azure.ServiceBus;
 
-namespace Equinor.ProCoSys.PcsServiceBus.Receiver.Interfaces
+namespace Equinor.ProCoSys.PcsServiceBus.Receiver.Interfaces;
+
+public interface IPcsSubscriptionClients
 {
-    public interface IPcsSubscriptionClients
-    {
-        Task CloseAllAsync();
-        void RegisterPcsMessageHandler(Func<IPcsSubscriptionClient, Message, CancellationToken, Task> handler, MessageHandlerOptions options);
-        void UnregisterPcsMessageHandler();
-        int RenewLeaseInterval { get; }
-    }
+    Task CloseAllAsync();
+    void RegisterPcsMessageHandler(Func<IPcsSubscriptionClient, Message, CancellationToken, Task> handler, MessageHandlerOptions options);
+    void UnregisterPcsMessageHandler();
+    int RenewLeaseInterval { get; }
 }
