@@ -2,20 +2,19 @@
 using Equinor.ProCoSys.PcsServiceBus.Receiver.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Equinor.ProCoSys.PcsServiceBus.Receiver
+namespace Equinor.ProCoSys.PcsServiceBus.Receiver;
+
+public class ScopedBusReceiverServiceFactory : IBusReceiverServiceFactory
 {
-    public class ScopedBusReceiverServiceFactory : IBusReceiverServiceFactory
+    private readonly IServiceProvider _services;
+
+    public ScopedBusReceiverServiceFactory(IServiceProvider services) => _services = services;
+
+    public IBusReceiverService GetServiceInstance()
     {
-        private readonly IServiceProvider _services;
+        var scope = _services.CreateScope();
+        var busReceiverService = scope.ServiceProvider.GetRequiredService<IBusReceiverService>();
 
-        public ScopedBusReceiverServiceFactory(IServiceProvider services) => _services = services;
-
-        public IBusReceiverService GetServiceInstance()
-        {
-            var scope = _services.CreateScope();
-            var busReceiverService = scope.ServiceProvider.GetRequiredService<IBusReceiverService>();
-
-            return busReceiverService;
-        }
+        return busReceiverService;
     }
 }
