@@ -1,11 +1,10 @@
-﻿namespace Equinor.ProCoSys.PcsServiceBus.Topics
+﻿namespace Equinor.ProCoSys.PcsServiceBus.Topics;
+
+public class IpoTopic
 {
-    public class IpoTopic
-    {
-        public string Plant { get; set; }
-        public string InvitationGuid { get; set; }
-        public string Event { get; set; }
-        public const string TopicName = "ipo";
-        public int Status { get; set; }
-    }
+    public string Plant { get; set; }
+    public string InvitationGuid { get; set; }
+    public string Event { get; set; }
+    public const string TopicName = "ipo";
+    public int Status { get; set; }
 }
