@@ -1,14 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Equinor.ProCoSys.PcsServiceBus.Topics
-{
-    public class McPkgDeleteTopic
-    {
-        public string Plant { get; set; }
-        public string McPkgNo { get; set; }
-        public string McPkgId { get; set; }
+namespace Equinor.ProCoSys.PcsServiceBus.Topics;
 
-        public const string TopicName = "mcpkg_delete";
-    }
+public class McPkgDeleteTopic
+{
+    public string Plant { get; set; }
+    public string McPkgNo { get; set; }
+    public string McPkgId { get; set; }
+
+    public const string TopicName = "mcpkg_delete";
 }
