@@ -1,12 +1,11 @@
-﻿namespace Equinor.ProCoSys.PcsServiceBus.Topics
-{
-    public class WoChecklistTopic
-    {
-        public string Plant { get; set; }
-        public string ProjectName { get; set; }
-        public string WoNo { get; set; }
-        public string ChecklistId { get; set; }
+﻿namespace Equinor.ProCoSys.PcsServiceBus.Topics;
 
-        public const string TopicName = "wochecklist";
-    }
+public class WoChecklistTopic
+{
+    public string Plant { get; set; }
+    public string ProjectName { get; set; }
+    public string WoNo { get; set; }
+    public string ChecklistId { get; set; }
+
+    public const string TopicName = "wochecklist";
 }
