@@ -1,12 +1,11 @@
 ﻿
-namespace Equinor.ProCoSys.PcsServiceBus.Topics
-{
-    public class WoMaterialDeleteTopic
-    {
-        public string Plant { get; set; }
-        public string WoNo { get; set; }
-        public string ItemNo { get; set; }
+namespace Equinor.ProCoSys.PcsServiceBus.Topics;
 
-        public const string TopicName = "womaterial_delete";
-    }
+public class WoMaterialDeleteTopic
+{
+    public string Plant { get; set; }
+    public string WoNo { get; set; }
+    public string ItemNo { get; set; }
+
+    public const string TopicName = "womaterial_delete";
 }
