@@ -2,17 +2,16 @@
 using Microsoft.EntityFrameworkCore;
 using Moq;
 
-namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Tests
-{
-    public class ContextHelper
-    {
-        public ContextHelper()
-        {
-            DbOptions = new DbContextOptions<BusSenderServiceContext>();
-            ContextMock = new Mock<BusSenderServiceContext>(DbOptions);
-        }
+namespace Equinor.ProCoSys.BusSenderWorker.Infrastructure.Tests;
 
-        public DbContextOptions<BusSenderServiceContext> DbOptions { get; }
-        public Mock<BusSenderServiceContext> ContextMock { get; }
+public class ContextHelper
+{
+    public ContextHelper()
+    {
+        DbOptions = new DbContextOptions<BusSenderServiceContext>();
+        ContextMock = new Mock<BusSenderServiceContext>(DbOptions);
     }
+
+    public DbContextOptions<BusSenderServiceContext> DbOptions { get; }
+    public Mock<BusSenderServiceContext> ContextMock { get; }
 }
